Many network changes. Too much for a one-liner.
This brings a long-needed refactor separating Player from the initial login sequence, allowing Players to not be created until they are ready to be fully constructed. This solves a wide range of problems with plugins doing odd things in PlayerPreLoginEvent and crashing the server, and also makes the Player construction code more sane and easier to logic out.

It also brings a swathe of other changes, including moving large parts of the login sequence logic into dedicated handler classes, allowing significant reduction in the size of Player (this isn't done yet, so it will continue to shrink as this work continues)
- DataPacketReceiveEvent has had getPlayer() removed, and now has getSource() which returns a PlayerNetworkSession instance.
- DataPacketSendEvent has had getPlayer() removed, and now has getTarget() which returns a PlayerNetworkSession instance.
- DataPacketBroadcastEvent now has a getTargets() and setTargets() which return and accept PlayerNetworkSession[] respectively.
- PlayerCreationEvent is now only called when the Player entity is ready to be constructed, and also allows manipulating the player's constructor parameters in the PlayerParameters object (which allows non-hacky username overrides and fun things like that).

This is not complete yet and some things are broken (notably login timeouts and duplicate player kicking).